decomp: fix 65 matched C files that did not compile; text/BG naming batch
Two independent issues, both surfaced while verifying a rename.

1. Matched-C integrity (pre-existing, not caused by this batch)
-------------------------------------------------------------
`make audit` found 65 of 633 matched C files could not be compiled at all
(568 clean). All failed with agbcc `conflicting types`: a naked wrapper declared
`(void)` against the real prototype in unknown-functions.h, plus wrong return
types (void vs s32, void vs void*) and one wrong parameter type (u32 a vs u16 a).

Nothing caught this because matched C is never linked -- the Makefile's C_SRCS is
empty and the ROM links generated asm/matchings/*.s -- so a file could be counted
as matching while not compiling. The sample offender sub_0803114C is unchanged at
HEAD and unnamed, so this predates the pipeline rework.

- tools/decomp/audit_c_compiles.py (`make audit`): compiles every
  src/matched/*.c standalone through agbcc, honouring match-flags and
  match-compiler.
- tools/decomp/repair_naked_signatures.py (`make repair-signatures`): copies the
  prototype signature into a disagreeing definition. Deliberately narrow -- only
  a naked-anchored or own-top-level definition -- and only for files that
  actually fail, so parameter-name-only differences (27 files) are left alone.
  An earlier loose draft mis-matched call sites and would have rewritten 290.
- 63 auto-repaired + 2 hand-fixed. Every touched file re-verified with
  match_function.py: 63 kept, 0 reverted, so no match regressed.

Result: make audit 633/633 clean (was 568); make compare OK on a forced rebuild.

2. Naming and throughput
------------------------
- agent_packet.py --next was skipping every cheap bucket, because all small_clean
  / leaf_branch / large targets are parked seeds and the old logic only fell back
  to parked seeds after exhausting fresh ones. It returned a fresh 1552-byte
  high_reg_pressure function; leaf_branch was unreachable entirely. Difficulty
  bucket is now the primary key with parked-vs-fresh as a tiebreak inside it.
- symbols.py: renaming an already-applied symbol stranded the old readable name
  in source while symbols.h stopped defining it (a compile error). Superseded
  names are now kept in aliases[] and apply rewrites every alias.
- Named 9 text/BG functions (15 -> 23/633). Engine layout recovered from call
  sites: +0x90/+0x92 cursor x/y, +0x96 palette bank, +0x98 area width, +0xA2 line
  height, +0x9C size in words.

- sub_08031300 rewritten as clean semantic C and parked: same size 78/78, 22/78
  bytes matched, permuter 240s x 8 no score 0. Register-destination difference
  (retail keeps the param in r2), same family as sub_08061308. Notes record it as
  do-not-retry.

Lifecycle: UNKNOWN 84 | DECOMPILED 130 | UNDERSTOOD 6 | MATCHING 413.
Named 23/633. make compare: OK. make audit: 0 failures.

+++ src/matched/sub_0806FDD0.c
@@ -2,7 +2,7 @@
 
 // @ 0x0806fdd0
 __attribute__((naked))
-void *sub_0806FDD0(u32 a)
+void * sub_0806FDD0(u16 a)
 {
     asm(
         ".syntax unified\n"
